Refactor to have separate files for different plots and the benchmark results processing.
